[31-01-2013 18:30] Détails d'un article. Enchère. On peut pas faire plus clair.

+++ GLAVenteAuxEncheres-ejb/src/java/library/ArticleBeanInterface.java
@@ -16,6 +16,6 @@ import persistence.Article;
 public interface ArticleBeanInterface {
     public List<Article> getCriticalsArticles();
     public List<Article> getArticlesInPromotion();
-    public int getActualPrice(Article a);
+    public double getActualPrice(Article a);
     public List<Article> search(String keywords);
 }
